Update LIB_PATHS in the makefile.
We were seeing a hardfault in startup_stm32h743xx.s when
going into __libc_init_array.  It appears it was linking to
ARM versions of the libraries, instead of thumb versions.

With this change I see it turning on LEDs, and running some of the USB code.  It enumerates as a "DAPLink CMSIS-DAP" device as well.